# etc/api-docs.yaml
openapi: 3.0.1
info:
  title: Ships API
  description: The backend for the Ships Portal
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/v1/fetch:
    post:
      tags:
      - fetch-controller
      operationId: fetchContent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FetchRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /api/v1/database/upload:
    post:
      tags:
      - database-controller
      summary: Upload content into the Database.
      description: "The maximum file size is 100MB, the file will be analyzed and\
        \ inserted into the Database"
      operationId: uploadDatabaseContent
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - multipartFile
              type: object
              properties:
                multipartFile:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /api/v1/database/delete/{entityName}:
    delete:
      tags:
      - database-controller
      operationId: deleteDatabase
      parameters:
      - name: entityName
        in: path
        description: delete the content of a single table
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableDetailsResponse'
components:
  schemas:
    FetchRequest:
      type: object
      properties:
        apiKey:
          type: string
        mmsi:
          type: string
        begin:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          enum:
          - SUCCESS
          - FAILED
          - UNDEFINED
        info:
          type: string
        description:
          type: string
    SingeTableDetails:
      type: object
      properties:
        name:
          type: string
        rowCount:
          type: integer
          format: int64
    TableDetailsResponse:
      type: object
      properties:
        tableDetails:
          type: array
          items:
            $ref: '#/components/schemas/SingeTableDetails'